Change byteswritten to uint32_t

+++ steim/csteim.pxd
@@ -1,8 +1,8 @@
-from libc.stdint cimport int32_t, int64_t, uint16_t
+from libc.stdint cimport int32_t, int64_t, uint32_t
 
 cdef extern from "packdata.h":
     extern int msr_encode_steim2(int32_t *input, int samplecount, int32_t *output,
-                                 int outputlength, int32_t diff0, uint16_t *byteswritten,
+                                 int outputlength, int32_t diff0, uint32_t *byteswritten,
                                  char *sid, int swapflag) nogil
 
 cdef extern from "unpackdata.h":
